fix; repository, controller, service
fix: duplicate user problem

+++ src/main/java/dev/clinican/dto/DoctorDto.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 public record DoctorDto(
         UUID id,
         UUID userId,
-        Long crm,
+        Integer crm,
         String speciality,
         Integer phoneNumber
 ) {
